13 예외 처리

+++ src/main/java/com/sparta/springbasicsolo/service/TodoService.java
@@ -1,6 +1,8 @@
 package com.sparta.springbasicsolo.service;
 
 import com.sparta.springbasicsolo.TodoDTO;
+import com.sparta.springbasicsolo.exception.DeletedTodoException;
+import com.sparta.springbasicsolo.exception.PasswordNotMatchedException;
 import com.sparta.springbasicsolo.repository.JdbcTodoRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +24,13 @@ public class TodoService {
     }
 
     public Optional<TodoDTO> findById(Long id) {
-        return todoRepository.findById(id);
+        Optional<TodoDTO> todoDTO = todoRepository.findById(id);
+        if (todoDTO.isPresent()) {
+            if (todoDTO.get().getIsDeleted()) {
+                throw new DeletedTodoException("이미 삭제된 일정입니다.");
+            }
+        }
+        return todoDTO;
     }
 
     public List<TodoDTO> findAll() {
@@ -38,7 +46,8 @@ public class TodoService {
                 if (updatedRowCount == 1) {
                     return todoRepository.findById(todoDTO.getId());
                 }
-                return Optional.empty();
+            } else {
+                throw new PasswordNotMatchedException("비밀번호가 일치하지 않습니다.");
             }
         }
         return Optional.empty();
@@ -47,9 +56,14 @@ public class TodoService {
     public void deleteTodo(Long id, String password) {
         Optional<TodoDTO> findTodo = todoRepository.findById(id);
         if (findTodo.isPresent()) {
+            if (findTodo.get().getIsDeleted()) {
+                throw new DeletedTodoException("이미 삭제된 일정입니다.");
+            }
             if (Objects.equals(findTodo.get().getPassword(), password)) {
                 log.info("비밀번호 일치함. 삭제 진행");
                 todoRepository.deleteTodo(id);
+            } else {
+                throw new PasswordNotMatchedException("비밀번호가 일치하지 않습니다.");
             }
         }
     }
